manager classes segregation
se separan los managers de experience y Job
