+++ app/src/androidTest/java/com/scorradi/cv/datamanager/DataManagerTest.kt
@@ -2,7 +2,6 @@ package com.scorradi.cv.datamanager
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import com.scorradi.cv.DataManager.DataManager
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -14,7 +13,7 @@ class DataManagerTest {
         val dataManager = DataManager()
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
 
-        val experiencesJson = dataManager.getExperiencesFromJson("experiences.json", appContext)
+        val experiencesJson = dataManager.getExperiences(appContext)
         val experiencesMocked = dataManager.getExperiencesMocked()
 
         Assert.assertEquals(experiencesMocked, experiencesJson)
