+++ app/src/main/java/com/scorradi/cv/datamanager/Utils.kt
@@ -1,0 +1,26 @@
+package com.scorradi.cv.datamanager
+
+import android.content.Context
+import java.io.IOException
+
+class Utils{
+    companion object {
+
+        fun loadJsonFromFile(filename: String, context: Context): String {
+            var json = ""
+
+            try {
+                val input = context.assets.open(filename)
+                val size = input.available()
+                val buffer = ByteArray(size)
+                input.read(buffer)
+                input.close()
+                json = buffer.toString(Charsets.UTF_8)
+            } catch (e: IOException) {
+                e.printStackTrace()
+            }
+
+            return json
+        }
+    }
+}
