+++ app/src/test/java/com/scorradi/cv/interactors/dataInteractorTest.kt
@@ -1,6 +1,5 @@
 package com.scorradi.cv.interactors
-import com.scorradi.cv.DataManager.DataManager
-import com.scorradi.cv.db.daos.entities.Person
+import com.scorradi.cv.datamanager.DataManager
 import com.scorradi.cv.views.models.PersonModel
 import org.junit.Test
